+++ include/cross2d/skeleton/audio.h
@@ -15,7 +15,7 @@ namespace c2d {
 
         typedef void (*C2DAudioCallback)(void *data, unsigned char *stream, int len);
 
-        Audio(int rate = 48000, int fps = 60, C2DAudioCallback cb = nullptr);
+        Audio(int rate = 48000, float fps = 60, C2DAudioCallback cb = nullptr);
 
         virtual ~Audio();
